Moved functionality in SurfaceAccess.get_realization_surface_intersection_async() into separate worker function
